Recordings were getting squashed. Edgecast would save the recording as streamName.mp4 whenever there was a conflict they would change it to streamName.1.mp4, .2, .3, etc
however since we are moving the files into a folder, there are no more conflicts
so all the files are streamName.mp4 (or maybe .1.mp4).
Point is we need to see how many streams currently exist in the folder with that name
and then set the correct stream .N suffix.