added account billing_plan and billing_provider models
Here's the structure:

Account is the master (umbrella) object.
Users can belong to multiple accounts
An Account has a single BillingProvider
A BillingProvider represents what source created the account (heroku/appdirect/cine.io/etc.)
A BillingPlan belongs to a BillingProvider
BillingPlans specify the limitations of each plan and the cost
An Account can have many BillingPlans. (This is a requirement of appdirect)
Projects belong to an Account